profile/esp/central/src/androidTest/java/org/im97mori/ble/profile/esp/central/db/EnvironmentalSensingProfileBondedDatabaseHelperTest.java
package org.im97mori.ble.profile.esp.central.db;

import androidx.test.core.app.ApplicationProvider;

import org.junit.Test;

import static org.junit.Assert.assertEquals;
import static org.junit.Assert.assertNotNull;

public class EnvironmentalSensingProfileBondedDatabaseHelperTest {

    @Test
    public void test_constructor_00001() {
        assertNotNull(EnvironmentalSensingProfileBondedDatabaseHelper.getInstance(ApplicationProvider.getApplicationContext()));
    }

    @Test
    public void test_constructor_00002() {
        assertEquals(EnvironmentalSensingProfileBondedDatabaseHelper.getInstance(ApplicationProvider.getApplicationContext()), EnvironmentalSensingProfileBondedDatabaseHelper.getInstance(ApplicationProvider.getApplicationContext()));
    }

    @Test
    public void test_getProfileName_00001() {
        assertEquals(EnvironmentalSensingProfileBondedDatabaseHelper.PROFILE_NAME, EnvironmentalSensingProfileBondedDatabaseHelper.getInstance(ApplicationProvider.getApplicationContext()).getProfileName());
    }
}
